Renamed directories accroding to comment
