+++ homework-03/cmd/gosearch/gosearh.go
@@ -4,23 +4,20 @@ package main
 import (
 	"flag"
 	"fmt"
-	"go-course-4/homework-02/pkg/crawler"
-	"go-course-4/homework-02/pkg/crawler/spider"
-	"go-course-4/homework-02/pkg/index"
+	"go-course-4/homework-03/pkg/crawler"
+	"go-course-4/homework-03/pkg/crawler/spider"
+	"go-course-4/homework-03/pkg/index"
 	"strings"
 )
 
 const (
-	depth   = 1
-	flagMsg = "Use parameter -s and add a preferable key word (-s go)"
-	errMsg  = "Someting went wrong"
-	unsMsg  = "Have not found any documents according to your key word"
+	depth = 1
 )
 
 func main() {
 	urls := []string{"https://golang.org", "https://www.practical-go-lessons.com/"}
 
-	sFlag := flag.String("s", "", flagMsg)
+	sFlag := flag.String("s", "", "Use parameter -s and add a preferable key word (-s go)")
 	flag.Parse()
 
 	if len(*sFlag) == 0 {
@@ -36,7 +33,7 @@ func main() {
 	for _, u := range urls {
 		links, err := s.Scan(u, depth)
 		if err != nil {
-			fmt.Println(errMsg)
+			fmt.Println("Someting went wrong")
 		}
 
 		for _, l := range links {
@@ -50,9 +47,11 @@ func main() {
 
 	idx := index.Ids(strings.ToLower(*sFlag))
 	if len(idx) == 0 {
-		fmt.Println(unsMsg)
+		fmt.Println("Have not found any documents according to your key word")
+		return
 	}
 
+	fmt.Println("test")
 	for _, i := range idx {
 		min, max := docs[0].ID, docs[len(docs)-1].ID
 		for min <= max {
