+++ homework-03/pkg/crawler/membot/membot.go
@@ -1,7 +1,7 @@
 package membot
 
 import (
-	"go-course-4/homework-02/pkg/crawler"
+	"go-course-4/homework-03/pkg/crawler"
 )
 
 // Service - имитация служба поискового робота.
