+++ homework-03/pkg/crawler/spider/spider.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"strings"
 
-	"go-course-4/homework-02/pkg/crawler"
+	"go-course-4/homework-03/pkg/crawler"
 
 	"golang.org/x/net/html"
 )
